+++ NEWS.md
@@ -1,3 +1,8 @@
+# rintcal 1.4.2
+* added more detail regarding and references to available calibration curves in the documentation for `ccurve`.
+* the `ccurve` code is now shorter and less repetitive.
+* new internal function (taken from the `rice` package) to deal more robustly with `BCAD` (and its `zero` option).
+
 # rintcal 1.4.1
 * `intcal.data` now plots D14C correctly when `BCAD=TRUE`.
 * `intcal.data` no longer truncates the second calibration curve if activated (e.g., `cc2=2`).